requirements.txt
adium-theme-ubuntu==0.3.4
amqp==2.2.1
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
Babel==2.5.0
bcrypt==3.1.3
billiard==3.5.0.3
celery==4.1.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.0.3
deployer==0.3.10
Django==1.11.4
django-celery-results==1.0.1
django-mongodb-engine==0.6.0
djangorestframework==3.6.3
djangotoolbox==1.8.0
docopt==0.6.1
enum34==1.1.6
futures==3.1.1
gyp==0.1
httplib2==0.10.3
hyperlink==17.3.1
idna==2.5
incremental==17.5.0
ipaddress==1.0.18
kombu==4.1.0
paramiko==2.2.1
pbr==3.1.1
pexpect==3.0
prettytable==0.7.2
prompt-toolkit==1.0.14
py4j==0.10.4
pyasn1==0.3.2
pycparser==2.18
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
pyliblzma==0.5.3
pymongo==3.5.0
PyNaCl==1.1.2
pyspark==2.2.0
pysqlite==1.0.1
python-sjsclient==0.7.0
python-swiftclient==3.3.0
pytz==2017.2
requests==2.14.2
rpm-python==4.12.0.1
setproctitle==1.1.8
six==1.10.0
SoftLayer==5.2.6
softlayer-object-storage==0.5.4
termcolor==1.1.0
Twisted==17.5.0
typing==3.6.1
unity-lens-photos==1.0
urlgrabber==3.9.1
urllib3==1.21.1
vine==1.1.4
wcwidth==0.1.7
yum-metadata-parser==1.1.4
zope.interface==4.4.2
